# docs/openapi.yaml
openapi: 3.0.3
info:
  title: 政务 WAF 控制面 API
  description: |
    政务 WAF 安全防护平台 — Phase 1 控制面后端。
    基于 Coraza + OWASP CRS 架构愿景；本阶段仅提供站点/节点/策略/规则/认证/审计等管理 API。
  version: 1.0.0-phase1
servers:
  - url: http://localhost:8080
    description: Local
paths:
...
  /api/v1/sites:
    get:
      tags: [Sites]
      summary: List sites
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/page_size'
      responses:
        '200': { description: OK }
    post:
      tags: [Sites]
      summary: Create site
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SiteCreate'
      responses:
        '200': { description: Created }
  /api/v1/sites/{id}:
    parameters:
      - $ref: '#/components/parameters/id'
    get:
      tags: [Sites]
      summary: Get site
      responses:
        '200': { description: OK }
    put:
      tags: [Sites]
      summary: Update site
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SiteUpdate'
      responses:
        '200': { description: OK }
    delete:
      tags: [Sites]
      summary: Delete site
      responses:
        '200': { description: OK }
...
  /api/v1/rules:
    get:
      tags: [Rules]
      summary: List rules
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/page_size'
      responses:
        '200': { description: OK }
    post:
      tags: [Rules]
      summary: Create rule
      responses:
        '200': { description: Created }
  /api/v1/rules/{id}:
    parameters:
      - $ref: '#/components/parameters/id'
    get:
      tags: [Rules]
      summary: Get rule
      responses:
        '200': { description: OK }
    put:
      tags: [Rules]
      summary: Update rule
      responses:
        '200': { description: OK }
    delete:
      tags: [Rules]
      summary: Delete rule
      responses:
        '200': { description: OK }
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    id:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    page:
      name: page
      in: query
      schema: { type: integer, default: 1 }
    page_size:
      name: page_size
      in: query
      schema: { type: integer, default: 20, maximum: 100 }
  schemas:
...
    SiteCreate:
      type: object
      required: [name, domain, upstream_port, protocol]
      properties:
        name: { type: string }
        domain: { type: string }
        upstream_host: { type: string }
        upstream_port: { type: integer, minimum: 1, maximum: 65535 }
        protocol: { type: string, enum: [http, https] }
        protection_mode: { type: string, enum: [observe, protect] }
        policy_id: { type: string, format: uuid, nullable: true }
        node_group_id: { type: string, format: uuid, nullable: true }
        status: { type: string, enum: [enabled, disabled] }
    SiteUpdate:
      type: object
      properties:
        name: { type: string }
        domain: { type: string }
        upstream_host: { type: string }
        upstream_port: { type: integer }
        protocol: { type: string, enum: [http, https] }
        protection_mode: { type: string, enum: [observe, protect] }
        policy_id: { type: string, nullable: true }
        node_group_id: { type: string, nullable: true }
        status: { type: string, enum: [enabled, disabled] }
security:
  - BearerAuth: []